requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
python-dateutil==2.8.2
six==1.16.0
openai==1.59.8
httpx==0.24.1
sambanova==0.3.1
flask-cors==4.0.0
pytz==2024.1
jsonschema==4.19.0
python-dotenv>=1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
PyJWT==2.8.0

